[INFRASTRUCTURE] Migrate to json based config. Move gemm test to integration. (#28)
* Migrate to json based config. Move gemm test to integration.

* temp checkin

* checkin  example json
