+++ vta/make/README.md
@@ -1,0 +1,8 @@
+# VTA Configuration
+
+Each VTA runtime/hardware configuration is specified by config.json file.
+You can copy the config.json to project root and modify the configuration
+before you type make.
+
+The config is going to affect the behavior of python package as well as
+the hardware runtime build.
