+++ vta/python/vta/__init__.py
@@ -1,15 +1,13 @@
 """TVM-based VTA Compiler Toolchain"""
 from __future__ import absolute_import as _abs
-import sys
+
 
 from .environment import get_env, Environment
+from . import arm_conv2d, vta_conv2d
+from .build_module import build_config, lower, build
+from .rpc_client import reconfig_runtime, program_fpga
 
 try:
-    # allow optional import in config mode.
-    from . import arm_conv2d, vta_conv2d
-    from .build_module import build_config, lower, build
-    from .rpc_client import reconfig_runtime, program_fpga
-
     from . import graph
 except (ImportError, RuntimeError):
     pass
